CreateTimeComponent: Added Time Component
