+++ src/components/CalculationPanel/CalculationPanel.module.css
@@ -14,18 +14,4 @@
 		align-items: center;
 		gap: 2rem;
 	}
-
-	.timer {
-		grid-area: timer;
-		display: flex;
-		flex-direction: row;
-		justify-content: center;
-		gap: 0.5rem;
-
-		font-weight: 500;
-
-		.time {
-			font-variant-numeric: tabular-nums;
-		}
-	}
 }
